remove cie cyclic dependencies
